refactor(executor): Split simd_ops/aggregate.rs into submodules
Split the 1322-line aggregate.rs file into focused submodules:
- sum.rs (319 lines) - SUM aggregate operations
- count.rs (79 lines) - COUNT aggregate operations
- minmax.rs (521 lines) - MIN/MAX aggregate operations
- product.rs (471 lines) - Product/arithmetic operations (e.g., SUM(a*b))
- mod.rs (60 lines) - Re-exports for backwards compatibility

Each module contains both basic and filtered variants (boolean and packed mask).
All existing tests pass. No changes to public API.

Closes #3291
