--- crates/vibesql-executor/src/select/columnar/simd_ops/aggregate.rs
@@ -1,1322 +1,0 @@
-//! Aggregate operations for columnar data processing.
-//!
-//! This module provides vectorized aggregate operations (sum, min, max, count)
-//! with both filtered and non-filtered variants using packed and boolean masks.
-
-#![allow(clippy::needless_range_loop)]
-
-use super::PackedMask;
-
-// ============================================================================
-// AGGREGATION OPERATIONS
-// ============================================================================
-
-/// Sum of i64 values using 4-accumulator auto-vectorization pattern.
-///
-/// Performance: Matches explicit SIMD (~2ms for 10M elements).
-#[inline]
-pub fn sum_i64(values: &[i64]) -> i64 {
-    let (mut s0, mut s1, mut s2, mut s3) = (0i64, 0i64, 0i64, 0i64);
-    let chunks = values.len() / 4;
-
-    for i in 0..chunks {
-        let off = i * 4;
-        s0 = s0.wrapping_add(values[off]);
-        s1 = s1.wrapping_add(values[off + 1]);
-        s2 = s2.wrapping_add(values[off + 2]);
-        s3 = s3.wrapping_add(values[off + 3]);
-    }
-
-    let mut sum = s0.wrapping_add(s1).wrapping_add(s2).wrapping_add(s3);
-    for i in (chunks * 4)..values.len() {
-        sum = sum.wrapping_add(values[i]);
-    }
-    sum
-}
-
-/// Sum of f64 values using 4-accumulator auto-vectorization pattern.
-///
-/// Performance: Matches explicit SIMD (~2ms for 10M elements).
-/// WARNING: Do NOT replace with `.iter().sum()` - it's 4x slower!
-#[inline]
-pub fn sum_f64(values: &[f64]) -> f64 {
-    let (mut s0, mut s1, mut s2, mut s3) = (0.0f64, 0.0f64, 0.0f64, 0.0f64);
-    let chunks = values.len() / 4;
-
-    for i in 0..chunks {
-        let off = i * 4;
-        s0 += values[off];
-        s1 += values[off + 1];
-        s2 += values[off + 2];
-        s3 += values[off + 3];
-    }
-
-    let mut sum = s0 + s1 + s2 + s3;
-    for i in (chunks * 4)..values.len() {
-        sum += values[i];
-    }
-    sum
-}
-
-/// Minimum of i64 values using 4-lane parallel reduction.
-#[inline]
-pub fn min_i64(values: &[i64]) -> Option<i64> {
-    if values.is_empty() {
-        return None;
-    }
-
-    let (mut m0, mut m1, mut m2, mut m3) = (i64::MAX, i64::MAX, i64::MAX, i64::MAX);
-    let chunks = values.len() / 4;
-
-    for i in 0..chunks {
-        let off = i * 4;
-        m0 = m0.min(values[off]);
-        m1 = m1.min(values[off + 1]);
-        m2 = m2.min(values[off + 2]);
-        m3 = m3.min(values[off + 3]);
-    }
-
-    let mut result = m0.min(m1).min(m2).min(m3);
-    for i in (chunks * 4)..values.len() {
-        result = result.min(values[i]);
-    }
-    Some(result)
-}
-
-/// Minimum of f64 values using 4-lane parallel reduction.
-#[inline]
-pub fn min_f64(values: &[f64]) -> Option<f64> {
-    if values.is_empty() {
-        return None;
-    }
-
-    let (mut m0, mut m1, mut m2, mut m3) =
-        (f64::INFINITY, f64::INFINITY, f64::INFINITY, f64::INFINITY);
-    let chunks = values.len() / 4;
-
-    for i in 0..chunks {
-        let off = i * 4;
-        m0 = m0.min(values[off]);
-        m1 = m1.min(values[off + 1]);
-        m2 = m2.min(values[off + 2]);
-        m3 = m3.min(values[off + 3]);
-    }
-
-    let mut result = m0.min(m1).min(m2).min(m3);
-    for i in (chunks * 4)..values.len() {
-        result = result.min(values[i]);
-    }
-    Some(result)
-}
-
-/// Maximum of i64 values using 4-lane parallel reduction.
-#[inline]
-pub fn max_i64(values: &[i64]) -> Option<i64> {
-    if values.is_empty() {
-        return None;
-    }
-
-    let (mut m0, mut m1, mut m2, mut m3) = (i64::MIN, i64::MIN, i64::MIN, i64::MIN);
-    let chunks = values.len() / 4;
-
-    for i in 0..chunks {
-        let off = i * 4;
-        m0 = m0.max(values[off]);
-        m1 = m1.max(values[off + 1]);
-        m2 = m2.max(values[off + 2]);
-        m3 = m3.max(values[off + 3]);
-    }
-
-    let mut result = m0.max(m1).max(m2).max(m3);
-    for i in (chunks * 4)..values.len() {
-        result = result.max(values[i]);
-    }
-    Some(result)
-}
-
-/// Maximum of f64 values using 4-lane parallel reduction.
-#[inline]
-pub fn max_f64(values: &[f64]) -> Option<f64> {
-    if values.is_empty() {
-        return None;
-    }
-
-    let (mut m0, mut m1, mut m2, mut m3) = (
-        f64::NEG_INFINITY,
-        f64::NEG_INFINITY,
-        f64::NEG_INFINITY,
-        f64::NEG_INFINITY,
-    );
-    let chunks = values.len() / 4;
-
-    for i in 0..chunks {
-        let off = i * 4;
-        m0 = m0.max(values[off]);
-        m1 = m1.max(values[off + 1]);
-        m2 = m2.max(values[off + 2]);
-        m3 = m3.max(values[off + 3]);
-    }
-
-    let mut result = m0.max(m1).max(m2).max(m3);
-    for i in (chunks * 4)..values.len() {
-        result = result.max(values[i]);
-    }
-    Some(result)
-}
-
-// ============================================================================
-// ARITHMETIC OPERATIONS (for expression aggregates)
-// ============================================================================
-
-/// Sum of element-wise product of two f64 arrays using 4-accumulator pattern.
-///
-/// This is the core operation for TPC-H Q6: SUM(l_extendedprice * l_discount).
-///
-/// Performance: Uses LLVM auto-vectorization for SIMD acceleration.
-/// The 4-accumulator pattern enables parallel execution across SIMD lanes.
-///
-/// # Arguments
-/// * `a` - First array of values
-/// * `b` - Second array of values (must be same length as `a`)
-///
-/// # Returns
-/// Sum of element-wise products: Σ(a[i] * b[i])
-///
-/// # Panics
-/// Panics if arrays have different lengths (debug builds only).
-#[inline]
-pub fn sum_product_f64(a: &[f64], b: &[f64]) -> f64 {
-    debug_assert_eq!(a.len(), b.len(), "Arrays must have same length");
-
-    let len = a.len().min(b.len());
-    if len == 0 {
-        return 0.0;
-    }
-
-    // 4-accumulator pattern for LLVM auto-vectorization
-    let (mut s0, mut s1, mut s2, mut s3) = (0.0f64, 0.0f64, 0.0f64, 0.0f64);
-    let chunks = len / 4;
-
-    for i in 0..chunks {
-        let off = i * 4;
-        s0 += a[off] * b[off];
-        s1 += a[off + 1] * b[off + 1];
-        s2 += a[off + 2] * b[off + 2];
-        s3 += a[off + 3] * b[off + 3];
-    }
-
-    let mut sum = s0 + s1 + s2 + s3;
-    for i in (chunks * 4)..len {
-        sum += a[i] * b[i];
-    }
-    sum
-}
-
-/// Sum of element-wise product with null mask using 4-accumulator pattern.
-///
-/// Handles NULL values efficiently by checking the null bitmap rather than
-/// per-element Option checking. This enables SIMD auto-vectorization.
-///
-/// # Arguments
-/// * `a` - First array of values
-/// * `b` - Second array of values (must be same length as `a`)
-/// * `null_a` - Null bitmap for array `a` (true = null, false = valid)
-/// * `null_b` - Null bitmap for array `b` (true = null, false = valid)
-///
-/// # Returns
-/// (sum, count) - Sum of products for non-null pairs and count of valid pairs
-#[inline]
-pub fn sum_product_f64_masked(
-    a: &[f64],
-    b: &[f64],
-    null_a: Option<&[bool]>,
-    null_b: Option<&[bool]>,
-) -> (f64, i64) {
-    let len = a.len().min(b.len());
-    if len == 0 {
-        return (0.0, 0);
-    }
-
-    // Fast path: no nulls in either array
-    let no_nulls_a = null_a.is_none_or(|n| !n.iter().any(|&x| x));
-    let no_nulls_b = null_b.is_none_or(|n| !n.iter().any(|&x| x));
-
-    if no_nulls_a && no_nulls_b {
-        return (sum_product_f64(a, b), len as i64);
-    }
-
-    // Slow path with null handling - still use accumulator pattern where possible
-    let null_mask_a = null_a.unwrap_or(&[]);
-    let null_mask_b = null_b.unwrap_or(&[]);
-
-    let mut sum = 0.0f64;
-    let mut count = 0i64;
-
-    for i in 0..len {
-        let is_null_a = null_mask_a.get(i).copied().unwrap_or(false);
-        let is_null_b = null_mask_b.get(i).copied().unwrap_or(false);
-
-        if !is_null_a && !is_null_b {
-            sum += a[i] * b[i];
-            count += 1;
-        }
-    }
-
-    (sum, count)
-}
-
-// ============================================================================
-// FILTERED AGGREGATION OPERATIONS (for fused filter+aggregate)
-// ============================================================================
-// These operations aggregate data directly using a filter mask, avoiding
-// intermediate allocations from creating filtered batches.
-
-/// Sum of f64 values where filter_mask[i] == true, using 4-accumulator pattern.
-///
-/// This fuses filtering and aggregation to avoid intermediate allocations.
-/// For simple aggregates like Q6, this can provide 2-3x speedup.
-///
-/// # Arguments
-/// * `values` - Array of values to aggregate
-/// * `filter_mask` - Boolean mask (true = include in aggregate)
-///
-/// # Returns
-/// Sum of values where filter_mask is true
-#[inline]
-pub fn sum_f64_filtered(values: &[f64], filter_mask: &[bool]) -> f64 {
-    debug_assert_eq!(values.len(), filter_mask.len(), "Arrays must have same length");
-
-    let len = values.len().min(filter_mask.len());
-    if len == 0 {
-        return 0.0;
-    }
-
-    // Use 4-accumulator pattern with conditional adds
-    let (mut s0, mut s1, mut s2, mut s3) = (0.0f64, 0.0f64, 0.0f64, 0.0f64);
-    let chunks = len / 4;
-
-    for i in 0..chunks {
-        let off = i * 4;
-        // 4-accumulator pattern: reduces loop-carried dependencies
-        // The compiler may auto-vectorize these conditional adds
-        if filter_mask[off] { s0 += values[off]; }
-        if filter_mask[off + 1] { s1 += values[off + 1]; }
-        if filter_mask[off + 2] { s2 += values[off + 2]; }
-        if filter_mask[off + 3] { s3 += values[off + 3]; }
-    }
-
-    let mut sum = s0 + s1 + s2 + s3;
-    for i in (chunks * 4)..len {
-        if filter_mask[i] {
-            sum += values[i];
-        }
-    }
-    sum
-}
-
-/// Sum of i64 values where filter_mask[i] == true.
-#[inline]
-pub fn sum_i64_filtered(values: &[i64], filter_mask: &[bool]) -> i64 {
-    debug_assert_eq!(values.len(), filter_mask.len(), "Arrays must have same length");
-
-    let len = values.len().min(filter_mask.len());
-    if len == 0 {
-        return 0;
-    }
-
-    let (mut s0, mut s1, mut s2, mut s3) = (0i64, 0i64, 0i64, 0i64);
-    let chunks = len / 4;
-
-    for i in 0..chunks {
-        let off = i * 4;
-        if filter_mask[off] { s0 = s0.wrapping_add(values[off]); }
-        if filter_mask[off + 1] { s1 = s1.wrapping_add(values[off + 1]); }
-        if filter_mask[off + 2] { s2 = s2.wrapping_add(values[off + 2]); }
-        if filter_mask[off + 3] { s3 = s3.wrapping_add(values[off + 3]); }
-    }
-
-    let mut sum = s0.wrapping_add(s1).wrapping_add(s2).wrapping_add(s3);
-    for i in (chunks * 4)..len {
-        if filter_mask[i] {
-            sum = sum.wrapping_add(values[i]);
-        }
-    }
-    sum
-}
-
-/// Count of true values in filter mask (number of rows passing filter).
-#[inline]
-pub fn count_filtered(filter_mask: &[bool]) -> i64 {
-    let len = filter_mask.len();
-    if len == 0 {
-        return 0;
-    }
-
-    // Use 4-accumulator pattern for count
-    let (mut c0, mut c1, mut c2, mut c3) = (0i64, 0i64, 0i64, 0i64);
-    let chunks = len / 4;
-
-    for i in 0..chunks {
-        let off = i * 4;
-        c0 += filter_mask[off] as i64;
-        c1 += filter_mask[off + 1] as i64;
-        c2 += filter_mask[off + 2] as i64;
-        c3 += filter_mask[off + 3] as i64;
-    }
-
-    let mut count = c0 + c1 + c2 + c3;
-    for i in (chunks * 4)..len {
-        count += filter_mask[i] as i64;
-    }
-    count
-}
-
-/// Sum of element-wise product of two f64 arrays with filter mask.
-///
-/// This is the key operation for TPC-H Q6: SUM(l_extendedprice * l_discount)
-/// with WHERE clause filtering.
-///
-/// Fuses filtering and aggregation to avoid:
-/// 1. Creating intermediate filtered arrays
-/// 2. Multiple passes over the data
-///
-/// # Arguments
-/// * `a` - First array of values
-/// * `b` - Second array of values (must be same length as `a`)
-/// * `filter_mask` - Boolean mask (true = include in aggregate)
-///
-/// # Returns
-/// (sum, count) - Sum of products for filtered rows and count of passing rows
-#[inline]
-pub fn sum_product_f64_filtered(a: &[f64], b: &[f64], filter_mask: &[bool]) -> (f64, i64) {
-    debug_assert_eq!(a.len(), b.len(), "Arrays must have same length");
-    debug_assert_eq!(a.len(), filter_mask.len(), "Arrays and filter must have same length");
-
-    let len = a.len().min(b.len()).min(filter_mask.len());
-    if len == 0 {
-        return (0.0, 0);
-    }
-
-    // 4-accumulator pattern with filter
-    let (mut s0, mut s1, mut s2, mut s3) = (0.0f64, 0.0f64, 0.0f64, 0.0f64);
-    let (mut c0, mut c1, mut c2, mut c3) = (0i64, 0i64, 0i64, 0i64);
-    let chunks = len / 4;
-
-    for i in 0..chunks {
-        let off = i * 4;
-        if filter_mask[off] {
-            s0 += a[off] * b[off];
-            c0 += 1;
-        }
-        if filter_mask[off + 1] {
-            s1 += a[off + 1] * b[off + 1];
-            c1 += 1;
-        }
-        if filter_mask[off + 2] {
-            s2 += a[off + 2] * b[off + 2];
-            c2 += 1;
-        }
-        if filter_mask[off + 3] {
-            s3 += a[off + 3] * b[off + 3];
-            c3 += 1;
-        }
-    }
-
-    let mut sum = s0 + s1 + s2 + s3;
-    let mut count = c0 + c1 + c2 + c3;
-
-    for i in (chunks * 4)..len {
-        if filter_mask[i] {
-            sum += a[i] * b[i];
-            count += 1;
-        }
-    }
-
-    (sum, count)
-}
-
-/// Sum of element-wise product with filter mask AND null masks.
-///
-/// Full version that handles both filter predicates and NULL values.
-/// Used when columnar data has NULLs that also need to be excluded.
-///
-/// # Arguments
-/// * `a` - First array of values
-/// * `b` - Second array of values
-/// * `filter_mask` - Boolean mask from predicates (true = include)
-/// * `null_a` - Null bitmap for array `a` (true = null, false = valid)
-/// * `null_b` - Null bitmap for array `b` (true = null, false = valid)
-///
-/// # Returns
-/// (sum, count) - Sum of products and count of valid rows
-#[inline]
-pub fn sum_product_f64_filtered_masked(
-    a: &[f64],
-    b: &[f64],
-    filter_mask: &[bool],
-    null_a: Option<&[bool]>,
-    null_b: Option<&[bool]>,
-) -> (f64, i64) {
-    let len = a.len().min(b.len()).min(filter_mask.len());
-    if len == 0 {
-        return (0.0, 0);
-    }
-
-    // Fast path: no nulls
-    let no_nulls_a = null_a.is_none_or(|n| n.len() < len || !n[..len].iter().any(|&x| x));
-    let no_nulls_b = null_b.is_none_or(|n| n.len() < len || !n[..len].iter().any(|&x| x));
-
-    if no_nulls_a && no_nulls_b {
-        return sum_product_f64_filtered(a, b, filter_mask);
-    }
-
-    // Slow path with null handling
-    let null_mask_a = null_a.unwrap_or(&[]);
-    let null_mask_b = null_b.unwrap_or(&[]);
-
-    let mut sum = 0.0f64;
-    let mut count = 0i64;
-
-    for i in 0..len {
-        let is_null_a = null_mask_a.get(i).copied().unwrap_or(false);
-        let is_null_b = null_mask_b.get(i).copied().unwrap_or(false);
-
-        if filter_mask[i] && !is_null_a && !is_null_b {
-            sum += a[i] * b[i];
-            count += 1;
-        }
-    }
-
-    (sum, count)
-}
-
-/// Min of f64 values where filter_mask[i] == true.
-#[inline]
-pub fn min_f64_filtered(values: &[f64], filter_mask: &[bool]) -> Option<f64> {
-    debug_assert_eq!(values.len(), filter_mask.len(), "Arrays must have same length");
-
-    let len = values.len().min(filter_mask.len());
-    if len == 0 {
-        return None;
-    }
-
-    let (mut m0, mut m1, mut m2, mut m3) =
-        (f64::INFINITY, f64::INFINITY, f64::INFINITY, f64::INFINITY);
-    let chunks = len / 4;
-
-    for i in 0..chunks {
-        let off = i * 4;
-        if filter_mask[off] { m0 = m0.min(values[off]); }
-        if filter_mask[off + 1] { m1 = m1.min(values[off + 1]); }
-        if filter_mask[off + 2] { m2 = m2.min(values[off + 2]); }
-        if filter_mask[off + 3] { m3 = m3.min(values[off + 3]); }
-    }
-
-    let mut result = m0.min(m1).min(m2).min(m3);
-    for i in (chunks * 4)..len {
-        if filter_mask[i] {
-            result = result.min(values[i]);
-        }
-    }
-
-    if result == f64::INFINITY {
-        None
-    } else {
-        Some(result)
-    }
-}
-
-/// Max of f64 values where filter_mask[i] == true.
-#[inline]
-pub fn max_f64_filtered(values: &[f64], filter_mask: &[bool]) -> Option<f64> {
-    debug_assert_eq!(values.len(), filter_mask.len(), "Arrays must have same length");
-
-    let len = values.len().min(filter_mask.len());
-    if len == 0 {
-        return None;
-    }
-
-    let (mut m0, mut m1, mut m2, mut m3) = (
-        f64::NEG_INFINITY,
-        f64::NEG_INFINITY,
-        f64::NEG_INFINITY,
-        f64::NEG_INFINITY,
-    );
-    let chunks = len / 4;
-
-    for i in 0..chunks {
-        let off = i * 4;
-        if filter_mask[off] { m0 = m0.max(values[off]); }
-        if filter_mask[off + 1] { m1 = m1.max(values[off + 1]); }
-        if filter_mask[off + 2] { m2 = m2.max(values[off + 2]); }
-        if filter_mask[off + 3] { m3 = m3.max(values[off + 3]); }
-    }
-
-    let mut result = m0.max(m1).max(m2).max(m3);
-    for i in (chunks * 4)..len {
-        if filter_mask[i] {
-            result = result.max(values[i]);
-        }
-    }
-
-    if result == f64::NEG_INFINITY {
-        None
-    } else {
-        Some(result)
-    }
-}
-
-/// Min of i64 values where filter_mask[i] == true.
-#[inline]
-pub fn min_i64_filtered(values: &[i64], filter_mask: &[bool]) -> Option<i64> {
-    debug_assert_eq!(values.len(), filter_mask.len(), "Arrays must have same length");
-
-    let len = values.len().min(filter_mask.len());
-    if len == 0 {
-        return None;
-    }
-
-    let (mut m0, mut m1, mut m2, mut m3) = (i64::MAX, i64::MAX, i64::MAX, i64::MAX);
-    let chunks = len / 4;
-
-    for i in 0..chunks {
-        let off = i * 4;
-        if filter_mask[off] { m0 = m0.min(values[off]); }
-        if filter_mask[off + 1] { m1 = m1.min(values[off + 1]); }
-        if filter_mask[off + 2] { m2 = m2.min(values[off + 2]); }
-        if filter_mask[off + 3] { m3 = m3.min(values[off + 3]); }
-    }
-
-    let mut result = m0.min(m1).min(m2).min(m3);
-    for i in (chunks * 4)..len {
-        if filter_mask[i] {
-            result = result.min(values[i]);
-        }
-    }
-
-    if result == i64::MAX {
-        None
-    } else {
-        Some(result)
-    }
-}
-
-/// Max of i64 values where filter_mask[i] == true.
-#[inline]
-pub fn max_i64_filtered(values: &[i64], filter_mask: &[bool]) -> Option<i64> {
-    debug_assert_eq!(values.len(), filter_mask.len(), "Arrays must have same length");
-
-    let len = values.len().min(filter_mask.len());
-    if len == 0 {
-        return None;
-    }
-
-    let (mut m0, mut m1, mut m2, mut m3) = (i64::MIN, i64::MIN, i64::MIN, i64::MIN);
-    let chunks = len / 4;
-
-    for i in 0..chunks {
-        let off = i * 4;
-        if filter_mask[off] { m0 = m0.max(values[off]); }
-        if filter_mask[off + 1] { m1 = m1.max(values[off + 1]); }
-        if filter_mask[off + 2] { m2 = m2.max(values[off + 2]); }
-        if filter_mask[off + 3] { m3 = m3.max(values[off + 3]); }
-    }
-
-    let mut result = m0.max(m1).max(m2).max(m3);
-    for i in (chunks * 4)..len {
-        if filter_mask[i] {
-            result = result.max(values[i]);
-        }
-    }
-
-    if result == i64::MIN {
-        None
-    } else {
-        Some(result)
-    }
-}
-
-// ============================================================================
-// PACKED FILTERED AGGREGATION OPERATIONS
-// ============================================================================
-// These operations aggregate data directly using a packed filter mask,
-// providing both memory efficiency and SIMD-friendly access patterns.
-
-/// Sum of f64 values where the corresponding bit in filter_mask is set.
-#[inline]
-pub fn sum_f64_packed_filtered(values: &[f64], filter_mask: &PackedMask) -> f64 {
-    debug_assert_eq!(values.len(), filter_mask.len(), "Arrays must have same length");
-
-    let len = values.len();
-    if len == 0 {
-        return 0.0;
-    }
-
-    // Process word by word for better cache utilization
-    let (mut s0, mut s1, mut s2, mut s3) = (0.0f64, 0.0f64, 0.0f64, 0.0f64);
-    let words = filter_mask.words();
-
-    for (word_idx, &word) in words.iter().enumerate() {
-        if word == 0 {
-            continue; // Skip entirely zero words
-        }
-
-        let base = word_idx * 64;
-        let end = (base + 64).min(len);
-
-        // Process set bits in this word
-        let mut bits = word;
-        while bits != 0 {
-            let bit_pos = bits.trailing_zeros() as usize;
-            let idx = base + bit_pos;
-            if idx < end {
-                // Distribute across 4 accumulators based on bit position
-                match bit_pos % 4 {
-                    0 => s0 += values[idx],
-                    1 => s1 += values[idx],
-                    2 => s2 += values[idx],
-                    _ => s3 += values[idx],
-                }
-            }
-            bits &= bits - 1; // Clear lowest set bit
-        }
-    }
-
-    s0 + s1 + s2 + s3
-}
-
-/// Sum of i64 values where the corresponding bit in filter_mask is set.
-#[inline]
-pub fn sum_i64_packed_filtered(values: &[i64], filter_mask: &PackedMask) -> i64 {
-    debug_assert_eq!(values.len(), filter_mask.len(), "Arrays must have same length");
-
-    let len = values.len();
-    if len == 0 {
-        return 0;
-    }
-
-    let (mut s0, mut s1, mut s2, mut s3) = (0i64, 0i64, 0i64, 0i64);
-    let words = filter_mask.words();
-
-    for (word_idx, &word) in words.iter().enumerate() {
-        if word == 0 {
-            continue;
-        }
-
-        let base = word_idx * 64;
-        let end = (base + 64).min(len);
-
-        let mut bits = word;
-        while bits != 0 {
-            let bit_pos = bits.trailing_zeros() as usize;
-            let idx = base + bit_pos;
-            if idx < end {
-                match bit_pos % 4 {
-                    0 => s0 = s0.wrapping_add(values[idx]),
-                    1 => s1 = s1.wrapping_add(values[idx]),
-                    2 => s2 = s2.wrapping_add(values[idx]),
-                    _ => s3 = s3.wrapping_add(values[idx]),
-                }
-            }
-            bits &= bits - 1;
-        }
-    }
-
-    s0.wrapping_add(s1).wrapping_add(s2).wrapping_add(s3)
-}
-
-/// Count of set bits in filter mask (number of rows passing filter).
-///
-/// This is a convenience wrapper around PackedMask::count_ones().
-#[inline]
-pub fn count_packed_filtered(filter_mask: &PackedMask) -> i64 {
-    filter_mask.count_ones() as i64
-}
-
-/// Sum of element-wise product of two f64 arrays with packed filter mask.
-///
-/// This is the key operation for TPC-H Q6: SUM(l_extendedprice * l_discount)
-/// with WHERE clause filtering.
-#[inline]
-pub fn sum_product_f64_packed_filtered(a: &[f64], b: &[f64], filter_mask: &PackedMask) -> (f64, i64) {
-    debug_assert_eq!(a.len(), b.len(), "Arrays must have same length");
-    debug_assert_eq!(a.len(), filter_mask.len(), "Arrays and filter must have same length");
-
-    let len = a.len();
-    if len == 0 {
-        return (0.0, 0);
-    }
-
-    let (mut s0, mut s1, mut s2, mut s3) = (0.0f64, 0.0f64, 0.0f64, 0.0f64);
-    let mut count = 0i64;
-    let words = filter_mask.words();
-
-    for (word_idx, &word) in words.iter().enumerate() {
-        if word == 0 {
-            continue;
-        }
-
-        let base = word_idx * 64;
-        let end = (base + 64).min(len);
-
-        let mut bits = word;
-        while bits != 0 {
-            let bit_pos = bits.trailing_zeros() as usize;
-            let idx = base + bit_pos;
-            if idx < end {
-                let product = a[idx] * b[idx];
-                match bit_pos % 4 {
-                    0 => s0 += product,
-                    1 => s1 += product,
-                    2 => s2 += product,
-                    _ => s3 += product,
-                }
-                count += 1;
-            }
-            bits &= bits - 1;
-        }
-    }
-
-    (s0 + s1 + s2 + s3, count)
-}
-
-/// Min of f64 values where the corresponding bit in filter_mask is set.
-#[inline]
-pub fn min_f64_packed_filtered(values: &[f64], filter_mask: &PackedMask) -> Option<f64> {
-    debug_assert_eq!(values.len(), filter_mask.len(), "Arrays must have same length");
-
-    let len = values.len();
-    if len == 0 {
-        return None;
-    }
-
-    let mut result = f64::INFINITY;
-    let words = filter_mask.words();
-
-    for (word_idx, &word) in words.iter().enumerate() {
-        if word == 0 {
-            continue;
-        }
-
-        let base = word_idx * 64;
-        let end = (base + 64).min(len);
-
-        let mut bits = word;
-        while bits != 0 {
-            let bit_pos = bits.trailing_zeros() as usize;
-            let idx = base + bit_pos;
-            if idx < end {
-                result = result.min(values[idx]);
-            }
-            bits &= bits - 1;
-        }
-    }
-
-    if result == f64::INFINITY {
-        None
-    } else {
-        Some(result)
-    }
-}
-
-/// Max of f64 values where the corresponding bit in filter_mask is set.
-#[inline]
-pub fn max_f64_packed_filtered(values: &[f64], filter_mask: &PackedMask) -> Option<f64> {
-    debug_assert_eq!(values.len(), filter_mask.len(), "Arrays must have same length");
-
-    let len = values.len();
-    if len == 0 {
-        return None;
-    }
-
-    let mut result = f64::NEG_INFINITY;
-    let words = filter_mask.words();
-
-    for (word_idx, &word) in words.iter().enumerate() {
-        if word == 0 {
-            continue;
-        }
-
-        let base = word_idx * 64;
-        let end = (base + 64).min(len);
-
-        let mut bits = word;
-        while bits != 0 {
-            let bit_pos = bits.trailing_zeros() as usize;
-            let idx = base + bit_pos;
-            if idx < end {
-                result = result.max(values[idx]);
-            }
-            bits &= bits - 1;
-        }
-    }
-
-    if result == f64::NEG_INFINITY {
-        None
-    } else {
-        Some(result)
-    }
-}
-
-/// Min of i64 values where the corresponding bit in filter_mask is set.
-#[inline]
-pub fn min_i64_packed_filtered(values: &[i64], filter_mask: &PackedMask) -> Option<i64> {
-    debug_assert_eq!(values.len(), filter_mask.len(), "Arrays must have same length");
-
-    let len = values.len();
-    if len == 0 {
-        return None;
-    }
-
-    let mut result = i64::MAX;
-    let words = filter_mask.words();
-
-    for (word_idx, &word) in words.iter().enumerate() {
-        if word == 0 {
-            continue;
-        }
-
-        let base = word_idx * 64;
-        let end = (base + 64).min(len);
-
-        let mut bits = word;
-        while bits != 0 {
-            let bit_pos = bits.trailing_zeros() as usize;
-            let idx = base + bit_pos;
-            if idx < end {
-                result = result.min(values[idx]);
-            }
-            bits &= bits - 1;
-        }
-    }
-
-    if result == i64::MAX {
-        None
-    } else {
-        Some(result)
-    }
-}
-
-/// Max of i64 values where the corresponding bit in filter_mask is set.
-#[inline]
-pub fn max_i64_packed_filtered(values: &[i64], filter_mask: &PackedMask) -> Option<i64> {
-    debug_assert_eq!(values.len(), filter_mask.len(), "Arrays must have same length");
-
-    let len = values.len();
-    if len == 0 {
-        return None;
-    }
-
-    let mut result = i64::MIN;
-    let words = filter_mask.words();
-
-    for (word_idx, &word) in words.iter().enumerate() {
-        if word == 0 {
-            continue;
-        }
-
-        let base = word_idx * 64;
-        let end = (base + 64).min(len);
-
-        let mut bits = word;
-        while bits != 0 {
-            let bit_pos = bits.trailing_zeros() as usize;
-            let idx = base + bit_pos;
-            if idx < end {
-                result = result.max(values[idx]);
-            }
-            bits &= bits - 1;
-        }
-    }
-
-    if result == i64::MIN {
-        None
-    } else {
-        Some(result)
-    }
-}
-
-/// Sum of element-wise product with packed filter mask AND null masks.
-///
-/// Full version that handles both filter predicates and NULL values.
-#[inline]
-pub fn sum_product_f64_packed_filtered_masked(
-    a: &[f64],
-    b: &[f64],
-    filter_mask: &PackedMask,
-    null_a: Option<&[bool]>,
-    null_b: Option<&[bool]>,
-) -> (f64, i64) {
-    let len = a.len().min(b.len()).min(filter_mask.len());
-    if len == 0 {
-        return (0.0, 0);
-    }
-
-    // Fast path: no nulls
-    let no_nulls_a = null_a.is_none_or(|n| n.len() < len || !n[..len].iter().any(|&x| x));
-    let no_nulls_b = null_b.is_none_or(|n| n.len() < len || !n[..len].iter().any(|&x| x));
-
-    if no_nulls_a && no_nulls_b {
-        return sum_product_f64_packed_filtered(a, b, filter_mask);
-    }
-
-    // Slow path with null handling
-    let null_mask_a = null_a.unwrap_or(&[]);
-    let null_mask_b = null_b.unwrap_or(&[]);
-
-    let mut sum = 0.0f64;
-    let mut count = 0i64;
-    let words = filter_mask.words();
-
-    for (word_idx, &word) in words.iter().enumerate() {
-        if word == 0 {
-            continue;
-        }
-
-        let base = word_idx * 64;
-        let end = (base + 64).min(len);
-
-        let mut bits = word;
-        while bits != 0 {
-            let bit_pos = bits.trailing_zeros() as usize;
-            let idx = base + bit_pos;
-            if idx < end {
-                let is_null_a = null_mask_a.get(idx).copied().unwrap_or(false);
-                let is_null_b = null_mask_b.get(idx).copied().unwrap_or(false);
-
-                if !is_null_a && !is_null_b {
-                    sum += a[idx] * b[idx];
-                    count += 1;
-                }
-            }
-            bits &= bits - 1;
-        }
-    }
-
-    (sum, count)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_sum_i64() {
-        let values: Vec<i64> = (1..=100).collect();
-        assert_eq!(sum_i64(&values), 5050);
-        assert_eq!(sum_i64(&[]), 0);
-        assert_eq!(sum_i64(&[42]), 42);
-    }
-
-    #[test]
-    fn test_sum_f64() {
-        let values: Vec<f64> = (1..=100).map(|x| x as f64).collect();
-        assert!((sum_f64(&values) - 5050.0).abs() < 0.001);
-        assert_eq!(sum_f64(&[]), 0.0);
-    }
-
-    #[test]
-    fn test_min_max_i64() {
-        let values = vec![5, 2, 8, 1, 9, 3, 7, 4, 6];
-        assert_eq!(min_i64(&values), Some(1));
-        assert_eq!(max_i64(&values), Some(9));
-        assert_eq!(min_i64(&[]), None);
-        assert_eq!(max_i64(&[]), None);
-    }
-
-    #[test]
-    fn test_min_max_f64() {
-        let values = vec![5.0, 2.0, 8.0, 1.0, 9.0, 3.0, 7.0, 4.0, 6.0];
-        assert_eq!(min_f64(&values), Some(1.0));
-        assert_eq!(max_f64(&values), Some(9.0));
-        assert_eq!(min_f64(&[]), None);
-        assert_eq!(max_f64(&[]), None);
-    }
-
-    #[test]
-    fn test_remainder_handling() {
-        // Test with non-multiple-of-4 lengths
-        let values: Vec<i64> = (1..=7).collect(); // 7 elements
-        assert_eq!(sum_i64(&values), 28);
-        assert_eq!(min_i64(&values), Some(1));
-        assert_eq!(max_i64(&values), Some(7));
-
-        let values: Vec<i64> = (1..=5).collect(); // 5 elements
-        assert_eq!(sum_i64(&values), 15);
-    }
-
-    #[test]
-    fn test_sum_product_f64() {
-        // Basic test: [1, 2, 3, 4] * [2, 2, 2, 2] = 2 + 4 + 6 + 8 = 20
-        let a = vec![1.0, 2.0, 3.0, 4.0];
-        let b = vec![2.0, 2.0, 2.0, 2.0];
-        assert!((sum_product_f64(&a, &b) - 20.0).abs() < 0.001);
-
-        // Empty arrays
-        assert_eq!(sum_product_f64(&[], &[]), 0.0);
-
-        // Single element
-        assert!((sum_product_f64(&[5.0], &[3.0]) - 15.0).abs() < 0.001);
-
-        // Non-multiple-of-4 lengths (tests remainder handling)
-        let a = vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0];
-        let b = vec![1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0];
-        assert!((sum_product_f64(&a, &b) - 28.0).abs() < 0.001);
-
-        // TPC-H Q6 style: price * discount
-        let prices = vec![100.0, 200.0, 300.0, 400.0, 500.0, 600.0, 700.0, 800.0];
-        let discounts = vec![0.05, 0.06, 0.07, 0.05, 0.06, 0.07, 0.05, 0.06];
-        // Expected: 100*0.05 + 200*0.06 + 300*0.07 + 400*0.05 + 500*0.06 + 600*0.07 + 700*0.05 + 800*0.06
-        //         = 5 + 12 + 21 + 20 + 30 + 42 + 35 + 48 = 213
-        assert!((sum_product_f64(&prices, &discounts) - 213.0).abs() < 0.001);
-    }
-
-    #[test]
-    fn test_sum_product_f64_masked() {
-        // No nulls - fast path
-        let a = vec![1.0, 2.0, 3.0, 4.0];
-        let b = vec![2.0, 2.0, 2.0, 2.0];
-        let (sum, count) = sum_product_f64_masked(&a, &b, None, None);
-        assert!((sum - 20.0).abs() < 0.001);
-        assert_eq!(count, 4);
-
-        // With nulls in first array
-        let null_a = vec![false, true, false, false];
-        let (sum, count) = sum_product_f64_masked(&a, &b, Some(&null_a), None);
-        // Only indices 0, 2, 3 counted: 1*2 + 3*2 + 4*2 = 2 + 6 + 8 = 16
-        assert!((sum - 16.0).abs() < 0.001);
-        assert_eq!(count, 3);
-
-        // With nulls in second array
-        let null_b = vec![false, false, true, false];
-        let (sum, count) = sum_product_f64_masked(&a, &b, None, Some(&null_b));
-        // Only indices 0, 1, 3 counted: 1*2 + 2*2 + 4*2 = 2 + 4 + 8 = 14
-        assert!((sum - 14.0).abs() < 0.001);
-        assert_eq!(count, 3);
-
-        // With nulls in both arrays
-        let (sum, count) = sum_product_f64_masked(&a, &b, Some(&null_a), Some(&null_b));
-        // Only indices where both are not null: 0, 3 → 1*2 + 4*2 = 10
-        assert!((sum - 10.0).abs() < 0.001);
-        assert_eq!(count, 2);
-
-        // Empty arrays
-        let (sum, count) = sum_product_f64_masked(&[], &[], None, None);
-        assert_eq!(sum, 0.0);
-        assert_eq!(count, 0);
-    }
-
-    #[test]
-    fn test_sum_f64_filtered() {
-        let values = vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0];
-        let filter = vec![true, false, true, false, true, false, true, false];
-        // Sum only odd positions: 1 + 3 + 5 + 7 = 16
-        assert!((sum_f64_filtered(&values, &filter) - 16.0).abs() < 0.001);
-
-        // All true
-        let filter_all = vec![true; 8];
-        assert!((sum_f64_filtered(&values, &filter_all) - 36.0).abs() < 0.001);
-
-        // All false
-        let filter_none = vec![false; 8];
-        assert_eq!(sum_f64_filtered(&values, &filter_none), 0.0);
-
-        // Empty
-        assert_eq!(sum_f64_filtered(&[], &[]), 0.0);
-    }
-
-    #[test]
-    fn test_sum_i64_filtered() {
-        let values = vec![1i64, 2, 3, 4, 5, 6, 7, 8];
-        let filter = vec![true, false, true, false, true, false, true, false];
-        assert_eq!(sum_i64_filtered(&values, &filter), 16);
-
-        let filter_all = vec![true; 8];
-        assert_eq!(sum_i64_filtered(&values, &filter_all), 36);
-    }
-
-    #[test]
-    fn test_count_filtered() {
-        let filter = vec![true, false, true, false, true, false, true, false];
-        assert_eq!(count_filtered(&filter), 4);
-
-        let filter_all = vec![true; 8];
-        assert_eq!(count_filtered(&filter_all), 8);
-
-        let filter_none = vec![false; 8];
-        assert_eq!(count_filtered(&filter_none), 0);
-
-        assert_eq!(count_filtered(&[]), 0);
-    }
-
-    #[test]
-    fn test_sum_product_f64_filtered() {
-        // TPC-H Q6 scenario: SUM(price * discount) WHERE filter
-        let prices = vec![100.0, 200.0, 300.0, 400.0, 500.0, 600.0, 700.0, 800.0];
-        let discounts = vec![0.05, 0.06, 0.07, 0.05, 0.06, 0.07, 0.05, 0.06];
-        let filter = vec![true, false, true, false, true, false, true, false];
-
-        // Only indices 0, 2, 4, 6: 100*0.05 + 300*0.07 + 500*0.06 + 700*0.05
-        //                        = 5 + 21 + 30 + 35 = 91
-        let (sum, count) = sum_product_f64_filtered(&prices, &discounts, &filter);
-        assert!((sum - 91.0).abs() < 0.001);
-        assert_eq!(count, 4);
-
-        // All pass filter
-        let filter_all = vec![true; 8];
-        let (sum, count) = sum_product_f64_filtered(&prices, &discounts, &filter_all);
-        assert!((sum - 213.0).abs() < 0.001);  // Same as unfiltered sum_product test
-        assert_eq!(count, 8);
-
-        // None pass filter
-        let filter_none = vec![false; 8];
-        let (sum, count) = sum_product_f64_filtered(&prices, &discounts, &filter_none);
-        assert_eq!(sum, 0.0);
-        assert_eq!(count, 0);
-    }
-
-    #[test]
-    fn test_sum_product_f64_filtered_masked() {
-        let a = vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0];
-        let b = vec![2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0];
-        let filter = vec![true, true, true, true, false, false, false, false];
-        let null_a = vec![false, true, false, false, false, false, false, false];
-
-        // Filter passes 0-3, null excludes 1, so: 0,2,3 → 1*2 + 3*2 + 4*2 = 16
-        let (sum, count) = sum_product_f64_filtered_masked(&a, &b, &filter, Some(&null_a), None);
-        assert!((sum - 16.0).abs() < 0.001);
-        assert_eq!(count, 3);
-    }
-
-    #[test]
-    fn test_min_max_f64_filtered() {
-        let values = vec![5.0, 2.0, 8.0, 1.0, 9.0, 3.0, 7.0, 4.0];
-        let filter = vec![true, false, true, false, true, false, true, false];
-
-        // Only indices 0, 2, 4, 6: values 5, 8, 9, 7
-        assert_eq!(min_f64_filtered(&values, &filter), Some(5.0));
-        assert_eq!(max_f64_filtered(&values, &filter), Some(9.0));
-
-        // No rows pass filter
-        let filter_none = vec![false; 8];
-        assert_eq!(min_f64_filtered(&values, &filter_none), None);
-        assert_eq!(max_f64_filtered(&values, &filter_none), None);
-    }
-
-    #[test]
-    fn test_min_max_i64_filtered() {
-        let values = vec![5i64, 2, 8, 1, 9, 3, 7, 4];
-        let filter = vec![true, false, true, false, true, false, true, false];
-
-        assert_eq!(min_i64_filtered(&values, &filter), Some(5));
-        assert_eq!(max_i64_filtered(&values, &filter), Some(9));
-
-        let filter_none = vec![false; 8];
-        assert_eq!(min_i64_filtered(&values, &filter_none), None);
-        assert_eq!(max_i64_filtered(&values, &filter_none), None);
-    }
-
-    #[test]
-    fn test_filtered_remainder_handling() {
-        // Test with non-multiple-of-4 lengths
-        let values: Vec<f64> = (1..=7).map(|x| x as f64).collect();  // 7 elements
-        let filter = vec![true, false, true, false, true, false, true];  // 1, 3, 5, 7 = 16
-        assert!((sum_f64_filtered(&values, &filter) - 16.0).abs() < 0.001);
-        assert_eq!(count_filtered(&filter), 4);
-    }
-
-    #[test]
-    fn test_sum_f64_packed_filtered() {
-        let values = vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0];
-        let mut mask = PackedMask::new_all_clear(8);
-        for i in (0..8).step_by(2) {
-            mask.set(i, true);
-        }
-        // Sum of 1, 3, 5, 7 = 16
-        assert!((sum_f64_packed_filtered(&values, &mask) - 16.0).abs() < 0.001);
-
-        // All true
-        let mask_all = PackedMask::new_all_set(8);
-        assert!((sum_f64_packed_filtered(&values, &mask_all) - 36.0).abs() < 0.001);
-
-        // All false
-        let mask_none = PackedMask::new_all_clear(8);
-        assert_eq!(sum_f64_packed_filtered(&values, &mask_none), 0.0);
-    }
-
-    #[test]
-    fn test_sum_i64_packed_filtered() {
-        let values = vec![1i64, 2, 3, 4, 5, 6, 7, 8];
-        let mut mask = PackedMask::new_all_clear(8);
-        for i in (0..8).step_by(2) {
-            mask.set(i, true);
-        }
-        assert_eq!(sum_i64_packed_filtered(&values, &mask), 16);
-    }
-
-    #[test]
-    fn test_count_packed_filtered() {
-        let mut mask = PackedMask::new_all_clear(8);
-        for i in (0..8).step_by(2) {
-            mask.set(i, true);
-        }
-        assert_eq!(count_packed_filtered(&mask), 4);
-    }
-
-    #[test]
-    fn test_sum_product_f64_packed_filtered() {
-        let prices = vec![100.0, 200.0, 300.0, 400.0, 500.0, 600.0, 700.0, 800.0];
-        let discounts = vec![0.05, 0.06, 0.07, 0.05, 0.06, 0.07, 0.05, 0.06];
-
-        let mut filter = PackedMask::new_all_clear(8);
-        for i in (0..8).step_by(2) {
-            filter.set(i, true);
-        }
-
-        // Only indices 0, 2, 4, 6: 100*0.05 + 300*0.07 + 500*0.06 + 700*0.05
-        //                        = 5 + 21 + 30 + 35 = 91
-        let (sum, count) = sum_product_f64_packed_filtered(&prices, &discounts, &filter);
-        assert!((sum - 91.0).abs() < 0.001);
-        assert_eq!(count, 4);
-
-        // All pass filter
-        let filter_all = PackedMask::new_all_set(8);
-        let (sum, count) = sum_product_f64_packed_filtered(&prices, &discounts, &filter_all);
-        assert!((sum - 213.0).abs() < 0.001);
-        assert_eq!(count, 8);
-    }
-
-    #[test]
-    fn test_min_max_f64_packed_filtered() {
-        let values = vec![5.0, 2.0, 8.0, 1.0, 9.0, 3.0, 7.0, 4.0];
-        let mut filter = PackedMask::new_all_clear(8);
-        for i in (0..8).step_by(2) {
-            filter.set(i, true);
-        }
-
-        // Only indices 0, 2, 4, 6: values 5, 8, 9, 7
-        assert_eq!(min_f64_packed_filtered(&values, &filter), Some(5.0));
-        assert_eq!(max_f64_packed_filtered(&values, &filter), Some(9.0));
-
-        // No rows pass filter
-        let filter_none = PackedMask::new_all_clear(8);
-        assert_eq!(min_f64_packed_filtered(&values, &filter_none), None);
-        assert_eq!(max_f64_packed_filtered(&values, &filter_none), None);
-    }
-
-    #[test]
-    fn test_min_max_i64_packed_filtered() {
-        let values = vec![5i64, 2, 8, 1, 9, 3, 7, 4];
-        let mut filter = PackedMask::new_all_clear(8);
-        for i in (0..8).step_by(2) {
-            filter.set(i, true);
-        }
-
-        assert_eq!(min_i64_packed_filtered(&values, &filter), Some(5));
-        assert_eq!(max_i64_packed_filtered(&values, &filter), Some(9));
-
-        let filter_none = PackedMask::new_all_clear(8);
-        assert_eq!(min_i64_packed_filtered(&values, &filter_none), None);
-        assert_eq!(max_i64_packed_filtered(&values, &filter_none), None);
-    }
-}
